Move parsers to their own file because they don't rely on anything else.
